requirements.txt
fastmcp>=2.11.0
httpx>=0.25.2
python-dotenv>=1.0.0
